Fix problems caused by merging 'main'
